// frontend-react/src/generated/screening.ts
export interface StartScreeningResponse {
    /**
     * @generated from protobuf field: string screening_id = 1
     */
    screeningId: string;
    /**
     * @generated from protobuf field: bool success = 2
     */
    success: boolean;
    /**
     * @generated from protobuf field: string error_message = 3
     */
    errorMessage: string;
    /**
     * @generated from protobuf field: talentpilot.screening.Question first_question = 4
     */
    firstQuestion?: Question;
}
// ...
export interface SubmitAnswerResponse {
    /**
     * @generated from protobuf field: talentpilot.screening.AnswerAssessment assessment = 1
     */
    assessment?: AnswerAssessment;
    /**
     * @generated from protobuf field: talentpilot.screening.Question next_question = 2
     */
    nextQuestion?: Question;
    /**
     * @generated from protobuf field: bool is_complete = 3
     */
    isComplete: boolean;
    /**
     * @generated from protobuf field: talentpilot.screening.EmailDraft email_draft = 4
     */
    emailDraft?: EmailDraft; // Available when screening is complete
}
// ...
export interface GetScreeningResultResponse {
    /**
     * @generated from protobuf field: bool success = 1
     */
    success: boolean;
    /**
     * @generated from protobuf field: talentpilot.screening.ScreeningSummary summary = 2
     */
    summary?: ScreeningSummary;
    /**
     * @generated from protobuf field: talentpilot.screening.EmailDraft email_draft = 3
     */
    emailDraft?: EmailDraft;
    /**
     * @generated from protobuf field: repeated talentpilot.screening.QuestionAnswer qa_history = 4
     */
    qaHistory: QuestionAnswer[];
    /**
     * @generated from protobuf field: string error_message = 5
     */
    errorMessage: string;
}
// ...
export interface Question {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
    /**
     * @generated from protobuf field: string text = 2
     */
    text: string;
    /**
     * @generated from protobuf field: string type = 3
     */
    type: string; // TECHNICAL_DEPTH, EXPERIENCE_VERIFICATION, PROBLEM_SOLVING, CULTURE_FIT, GAP_PROBE
    /**
     * @generated from protobuf field: string focus_area = 4
     */
    focusArea: string;
    /**
     * @generated from protobuf field: repeated string expected_evidence = 5
     */
    expectedEvidence: string[];
    /**
     * @generated from protobuf field: string priority = 6
     */
    priority: string; // REQUIRED, IMPORTANT, BONUS
}
// ...
export interface AnswerAssessment {
    /**
     * @generated from protobuf field: string question_id = 1
     */
    questionId: string;
    /**
     * @generated from protobuf field: string quality = 2
     */
    quality: string;
    /**
     * @generated from protobuf field: float confidence = 3
     */
    confidence: number;
    /**
     * @generated from protobuf field: repeated string key_points_identified = 4
     */
    keyPointsIdentified: string[];
    /**
     * @generated from protobuf field: repeated string gaps_identified = 5
     */
    gapsIdentified: string[];
    /**
     * @generated from protobuf field: string decision = 6
     */
    decision: string;
    /**
     * @generated from protobuf field: string reasoning = 7
     */
    reasoning: string;
}
/**
 * @generated from protobuf message talentpilot.screening.EmailDraft
 */
export interface EmailDraft {
    /**
     * @generated from protobuf field: string to = 1
     */
    to: string;
    /**
     * @generated from protobuf field: string subject = 2
     */
    subject: string;
    /**
     * @generated from protobuf field: string body = 3
     */
    body: string;
    /**
     * @generated from protobuf field: string cc = 4
     */
    cc: string;
    /**
     * @generated from protobuf field: string bcc = 5
     */
    bcc: string;
}
// ...
export interface QuestionAnswer {
    /**
     * @generated from protobuf field: talentpilot.screening.Question question = 1
     */
    question?: Question;
    /**
     * @generated from protobuf field: string answer_text = 2
     */
    answerText: string;
    /**
     * @generated from protobuf field: talentpilot.screening.AnswerAssessment assessment = 3
     */
    assessment?: AnswerAssessment;
    /**
     * @generated from protobuf field: string response_time_seconds = 4
     */
    responseTimeSeconds: string;
    /**
     * @generated from protobuf field: string timestamp = 5
     */
    timestamp: string;
}
// ...
class StartScreeningResponse$Type extends MessageType<StartScreeningResponse> {
    constructor() {
        super("talentpilot.screening.StartScreeningResponse", [
            { no: 1, name: "screening_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "success", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 3, name: "error_message", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "first_question", kind: "message", T: () => Question }
        ]);
    }
    create(value?: PartialMessage<StartScreeningResponse>): StartScreeningResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.screeningId = "";
        message.success = false;
        message.errorMessage = "";
        if (value !== undefined)
            reflectionMergePartial<StartScreeningResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: StartScreeningResponse): StartScreeningResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string screening_id */ 1:
                    message.screeningId = reader.string();
                    break;
                case /* bool success */ 2:
                    message.success = reader.bool();
                    break;
                case /* string error_message */ 3:
                    message.errorMessage = reader.string();
                    break;
                case /* talentpilot.screening.Question first_question */ 4:
                    message.firstQuestion = Question.internalBinaryRead(reader, reader.uint32(), options, message.firstQuestion);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: StartScreeningResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string screening_id = 1; */
        if (message.screeningId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.screeningId);
        /* bool success = 2; */
        if (message.success !== false)
            writer.tag(2, WireType.Varint).bool(message.success);
        /* string error_message = 3; */
        if (message.errorMessage !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.errorMessage);
        /* talentpilot.screening.Question first_question = 4; */
        if (message.firstQuestion)
            Question.internalBinaryWrite(message.firstQuestion, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message talentpilot.screening.StartScreeningResponse
 */
export const StartScreeningResponse = new StartScreeningResponse$Type();
// ...
class SubmitAnswerResponse$Type extends MessageType<SubmitAnswerResponse> {
    constructor() {
        super("talentpilot.screening.SubmitAnswerResponse", [
            { no: 1, name: "assessment", kind: "message", T: () => AnswerAssessment },
            { no: 2, name: "next_question", kind: "message", T: () => Question },
            { no: 3, name: "is_complete", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "email_draft", kind: "message", T: () => EmailDraft }
        ]);
    }
    create(value?: PartialMessage<SubmitAnswerResponse>): SubmitAnswerResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.isComplete = false;
        if (value !== undefined)
            reflectionMergePartial<SubmitAnswerResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SubmitAnswerResponse): SubmitAnswerResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* talentpilot.screening.AnswerAssessment assessment */ 1:
                    message.assessment = AnswerAssessment.internalBinaryRead(reader, reader.uint32(), options, message.assessment);
                    break;
                case /* talentpilot.screening.Question next_question */ 2:
                    message.nextQuestion = Question.internalBinaryRead(reader, reader.uint32(), options, message.nextQuestion);
                    break;
                case /* bool is_complete */ 3:
                    message.isComplete = reader.bool();
                    break;
                case /* talentpilot.screening.EmailDraft email_draft */ 4:
                    message.emailDraft = EmailDraft.internalBinaryRead(reader, reader.uint32(), options, message.emailDraft);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SubmitAnswerResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* talentpilot.screening.AnswerAssessment assessment = 1; */
        if (message.assessment)
            AnswerAssessment.internalBinaryWrite(message.assessment, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* talentpilot.screening.Question next_question = 2; */
        if (message.nextQuestion)
            Question.internalBinaryWrite(message.nextQuestion, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* bool is_complete = 3; */
        if (message.isComplete !== false)
            writer.tag(3, WireType.Varint).bool(message.isComplete);
        /* talentpilot.screening.EmailDraft email_draft = 4; */
        if (message.emailDraft)
            EmailDraft.internalBinaryWrite(message.emailDraft, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message talentpilot.screening.SubmitAnswerResponse
 */
export const SubmitAnswerResponse = new SubmitAnswerResponse$Type();
// ...
export const GetScreeningResultResponse = new GetScreeningResultResponse$Type();
// ...
export const Question = new Question$Type();
// ...
export const AnswerAssessment = new AnswerAssessment$Type();
// ...
export const EmailDraft = new EmailDraft$Type();
// ...
export const ScreeningSummary = new ScreeningSummary$Type();
// ...
export const QuestionAnswer = new QuestionAnswer$Type();